Registering UsersStore in constructor instead of StoreConfig
